+++ src/tokenizer.js
@@ -8,10 +8,13 @@
 
 const crypto = require('crypto');
 
-const TOKEN_PATTERN = /\[([A-Z_]+_(?:\d+|REDACTED))\]/g;
-const ESCAPED_OPEN = '\uFF3B';
-const ESCAPED_CLOSE = '\uFF3D';
-const ESCAPED_PATTERN = new RegExp(`${ESCAPED_OPEN}([A-Z_]+_(?:\\d+|REDACTED))${ESCAPED_CLOSE}`, 'g');
+const {
+  CLOAKLLM_TOKEN_REGEX: TOKEN_PATTERN,
+  ESCAPED_OPEN,
+  ESCAPED_CLOSE,
+} = require('./token-spec');
+
+const ESCAPED_PATTERN = new RegExp(`${ESCAPED_OPEN}([A-Z][A-Z0-9_]*_(?:\\d+|REDACTED))${ESCAPED_CLOSE}`, 'g');
 
 class TokenMap {
   /**